Move exchange-constants and retriers to exchange.common
