+++ freqtrade/exchange/__init__.py
@@ -1,4 +1,5 @@
-from freqtrade.exchange.exchange import Exchange, MAP_EXCHANGE_CHILDCLASS  # noqa: F401
+from freqtrade.exchange.common import MAP_EXCHANGE_CHILDCLASS  # noqa: F401
+from freqtrade.exchange.exchange import Exchange  # noqa: F401
 from freqtrade.exchange.exchange import (get_exchange_bad_reason,  # noqa: F401
                                          is_exchange_bad,
                                          is_exchange_known_ccxt,
